requirements.txt
# Web API
fastapi==0.119.1
uvicorn[standard]==0.34.3
pydantic==2.12.5

# Agent stack
langchain==0.3.28
langchain-core==0.3.84
langchain-community==0.3.31
langchain-groq==0.2.5
langchain-mcp-adapters==0.1.9   # 0.2.x requires langchain-core>=1.0.0 which conflicts
langchain-huggingface==0.1.2
langgraph==0.5.4

# Memory / embeddings
faiss-cpu==1.13.2
sentence-transformers==5.4.0

# MCP servers and integrations
mcp==1.27.0
python-dotenv==1.2.2
google-auth-oauthlib==1.3.1
google-auth-httplib2==0.3.1
google-api-python-client==2.194.0
tavily-python==0.7.23

# Tool server runtime deps
numpy>=2.0.0,<3.0.0
sympy>=1.13.0,<2.0.0
scipy>=1.14.0,<2.0.0
requests>=2.32.0,<3.0.0
httpx>=0.27.0,<1.0.0